11/02

+++ ft_printf/libft/ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: skim <skim@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 17:50:02 by skim              #+#    #+#             */
-/*   Updated: 2020/11/02 18:02:20 by skim             ###   ########.fr       */
+/*   Updated: 2020/11/02 21:04:37 by skim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long long	ft_pow(unsigned long long num, int exp)
+ull		ft_pow(unsigned long long num, int exp)
 {
 	int	n;
